Create Codebook
This file includes detailed information on the variables included in the dataset.  Here you can find what numbers indicate in each variable.  The variables that are self-explanatory do not include a notation. For example, GPA does not need a coding system, because participants have entered their GPA in numeric form.